requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
Faker==33.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
pillow==11.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
